+++ CHANGELOG.md
@@ -25,10 +25,11 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 - Defaults to StoreKit 2 for product purchasing for apps running on iOS 15+. You can change this back to StoreKit 1 by setting the `SuperwallOption` `storeKitVersion` to `.storeKit1`.
 - Changes the `PurchaseController` purchase function to `func purchase(product: StoreProduct) async -> PurchaseResult`. There will be an StoreKit 2 product accessible via `product.sk2Product` by default. However, if you're using the StoreKit 1 `SuperwallOption` or your app is running on an iOS version lower than iOS 15, this will be `nil` and you can access the StoreKit 1 product via `product.sk1Product`.
 - Consumables no longer count as lifetime subscriptions when using StoreKit 2.
--
+- Changes `presented_by_event_name` to `presented_by_placement_name`.
 
 ### Enhancements
 
+- Adds `purchase(_:)` support for both StoreKit 2 products and `StoreProduct`.
 - Updates the RevenueCat example app to use StoreKit 2.
 
 ## 3.10.1